refactor (schedulers/condition): breakdown large file of scheduler into smaller ones
