+++ internal/schedulers/condition/scheduler/worker/dispatch.go
@@ -1,0 +1,32 @@
+package worker
+
+import "time"
+
+// performActionExecution handles the actual action execution logic
+func (w *ConditionWorker) performActionExecution(triggerValue float64) bool {
+	// TODO: Implement actual action execution logic
+	// This should:
+	// 1. Send task to manager/keeper for execution
+	// 2. Handle response and update job status
+	// 3. For non-recurring jobs, stop the worker
+
+	w.Logger.Info("Simulating action execution",
+		"job_id", w.Job.JobID,
+		"trigger_value", triggerValue,
+		"target_chain", w.Job.TargetChainID,
+		"target_contract", w.Job.TargetContractAddress,
+		"target_function", w.Job.TargetFunction,
+	)
+
+	// Simulate processing time
+	time.Sleep(100 * time.Millisecond)
+
+	// For non-recurring jobs, stop the worker after execution
+	if !w.Job.Recurring {
+		w.Logger.Info("Non-recurring job completed, stopping worker", "job_id", w.Job.JobID)
+		go w.Stop() // Stop in a goroutine to avoid deadlock
+	}
+
+	// Simulate 95% success rate
+	return time.Now().UnixNano()%100 < 95
+}
